# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation untuk API Manajemen Senjata",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/detail-pengiriman": {
            "get": {
                "tags": [
                    "DetailPengiriman"
                ],
                "summary": "Menampilkan semua detail pengiriman",
                "operationId": "61df19f560d29df7b20a509e805959eb",
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DetailPengiriman"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DetailPengiriman"
                ],
                "summary": "Menambahkan data detail pengiriman",
                "operationId": "8fee43dbacb52fbceb22805b3702439a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pengiriman",
                                    "id_senjata",
                                    "jumlah",
                                    "created_at"
                                ],
                                "properties": {
                                    "id_pengiriman": {
                                        "type": "integer"
                                    },
                                    "id_senjata": {
                                        "type": "integer"
                                    },
                                    "jumlah": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailPengiriman"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/detail-pengiriman/{id}": {
            "get": {
                "tags": [
                    "DetailPengiriman"
                ],
                "summary": "Menampilkan detail pengiriman berdasarkan ID",
                "operationId": "39f1c8f7c8d7c0ec0c4d4347dacb567d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailPengiriman"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DetailPengiriman"
                ],
                "summary": "Mengupdate data detail pengiriman",
                "operationId": "203098f7913d8db7b673a9d9fc42d660",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pengiriman",
                                    "id_senjata",
                                    "jumlah",
                                    "created_at"
                                ],
                                "properties": {
                                    "id_pengiriman": {
                                        "type": "integer"
                                    },
                                    "id_senjata": {
                                        "type": "integer"
                                    },
                                    "jumlah": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetailPengiriman"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DetailPengiriman"
                ],
                "summary": "Menghapus data detail pengiriman",
                "operationId": "69701dac8ca200dc0b51fd9651c1f6f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "DetailPengiriman": {
                "required": [
                    "id_pengiriman",
                    "id_senjata",
                    "jumlah",
                    "created_at"
                ],
                "properties": {
                    "id_detail_pengiriman": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_pengiriman": {
                        "type": "integer",
                        "example": 1001
                    },
                    "id_senjata": {
                        "type": "integer",
                        "example": 2002
                    },
                    "jumlah": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-15T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "sanctum": []
        }
    }
}